requirements.txt
scalecast
tensorflow-transform
tensorflow
scikit-learn
Jinja2
Flask
plotly
sklearn-learn
matplotlib
pandas
tensorflow
keras-tuner
requests
numpy
os
json
re
statistics
sys
